Add LogsInjectionHelper
